refactor(plugin): extract marketplace package transfer into its own module
`crates/backend/src/plugin.rs` carries a reviewed size exception whose split
plan asks for its responsibilities to be separated. Marketplace release
resolution and package transfer form a self-contained one: they are the only
callers of the release manifest lookup, the host-target selection, the proxy
and object-store installer construction, and the install/update error maps.

Moving them to `plugin/marketplace.rs` takes plugin.rs from 1053 to 840
production lines, so the baseline is lowered to keep the removed debt from
growing back. The code is moved verbatim; the module keeps access to the
private plugin-host state it uses because it is a descendant of the module
that declares it.
